Broke out claims module into new facade module.
claims/__init__.py was split between facade/__init__.py and
facade/f847.py.
